refactor: mapper 위치 이동
